- vectors are everywhere now
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@28 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinxstemen.cpp
@@ -177,7 +177,7 @@ void stem_en ( BYTE * word )
 {
 	int i, j, len, r1, r2;
 
-	len = strlen((char*)word);
+	len = (int)strlen((char*)word);
 	if (len <= 2) return;
 
 	for (i = 0; i < (int)(sizeof(en_except)/sizeof(stem_table)); i++) {
